requirements.txt
python-dotenv==1.0.1
fastapi==0.114.2
pydantic==2.9.1
pydantic_core==2.23.3
uvicorn==0.30.6
requests==2.32.3
typing_extensions==4.12.2
python-dotenv==1.0.1
